handle flattened items & multiple maps
reasonably ok first pass.

multiple people calling .next() while a current .next() is running
should not result in data loss, but could very well result in data
coming in out of order.

also previous _queued implementation was iterator based, which was
cool, where-as this one holds results in an queue, which isn't as
cool.